Move expression table generation to enrichment-model
